Move new components into components folder, set new routes
