+++ burnt-out-app/src/components/CourseRow.css
