+++ burnt-out-app/src/routes/profs.js
@@ -1,4 +1,4 @@
-import Table from '../components/Table';
+import Table from '../components-old/Table';
 
 export default function Profs() {
   return (
