--- burnt-out-app/src/routes/test-courses.js
@@ -1,8 +1,0 @@
-import React from "react";
-import CourseTable from "../components-v2/CourseTable";
-
-export default function TestCourses() {
-  return (
-    <CourseTable />
-  );
-}
